+++ nilearn/reporting/glm_reporter.py
@@ -17,18 +17,17 @@ from collections.abc import Iterable
 from decimal import Decimal
 from html import escape
 from pathlib import Path
-from string import Template
 
 import numpy as np
 import pandas as pd
 from matplotlib import pyplot as plt
 
-from nilearn import experimental
 from nilearn._utils import check_niimg, fill_doc
 from nilearn._utils.niimg import safe_get_data
-from nilearn._version import __version__
+from nilearn.experimental.reporting.glm_reporter import (
+    _make_surface_glm_report,
+)
 from nilearn.experimental.surface import SurfaceMasker
-from nilearn.externals import tempita
 from nilearn.maskers import NiftiMasker
 from nilearn.plotting import plot_glass_brain, plot_roi, plot_stat_map
 from nilearn.plotting.cm import _cmap_d as nilearn_cmaps
@@ -39,7 +38,7 @@ from nilearn.plotting.matrix_plotting import (
 )
 from nilearn.reporting.get_clusters_table import get_clusters_table
 from nilearn.reporting.html_report import HTMLReport
-from nilearn.reporting.utils import figure_to_png_base64, figure_to_svg_quoted
+from nilearn.reporting.utils import figure_to_svg_quoted
 
 with warnings.catch_warnings():
     warnings.simplefilter("ignore", FutureWarning)
@@ -286,213 +285,6 @@ def make_glm_report(
     # setting report size for better visual experience in Jupyter Notebooks.
     report_text.width, report_text.height = _check_report_dims(report_dims)
     return report_text
-
-
-def _make_surface_glm_report(
-    model,
-    contrasts=None,
-    title=None,
-    threshold=3.09,
-    alpha=0.001,
-    cluster_threshold=0,
-    height_control="fpr",
-):
-
-    title = f" - {title}" if title else ""
-
-    selected_attributes = [
-        "subject_label",
-        "drift_model",
-        "hrf_model",
-        "standardize",
-        "noise_model",
-        "t_r",
-        "target_shape",
-        "signal_scaling",
-        "scaling_axis",
-        "smoothing_fwhm",
-        "target_affine",
-        "slice_time_ref",
-    ]
-    attribute_units = {
-        "t_r": "seconds",
-        "high_pass": "Hertz",
-    }
-
-    if hasattr(model, "hrf_model") and model.hrf_model == "fir":
-        selected_attributes.append("fir_delays")
-
-    if hasattr(model, "drift_model"):
-        if model.drift_model == "cosine":
-            selected_attributes.append("high_pass")
-        elif model.drift_model == "polynomial":
-            selected_attributes.append("drift_order")
-
-    selected_attributes.sort()
-    parameters = {
-        attr_name: getattr(model, attr_name)
-        for attr_name in selected_attributes
-        if hasattr(model, attr_name)
-    }
-    for attribute_name_, attribute_unit_ in attribute_units.items():
-        if attribute_name_ in parameters:
-            parameters[f"{attribute_name_} ({attribute_unit_})"] = parameters[
-                attribute_name_
-            ]
-            parameters.pop(attribute_name_)
-
-    cluster_table_details = OrderedDict()
-    threshold = np.around(threshold, 3)
-    if height_control:
-        cluster_table_details.update({"Height control": height_control})
-        if alpha < 0.001:
-            alpha = f"{Decimal(alpha):.2E}"
-        cluster_table_details.update({"\u03B1": alpha})
-        cluster_table_details.update({"Threshold (computed)": threshold})
-    else:
-        cluster_table_details.update({"Height control": "None"})
-        cluster_table_details.update({"Threshold Z": threshold})
-    cluster_table_details.update(
-        {"Cluster size threshold (vertices)": cluster_threshold}
-    )
-
-    warning_messages = []
-    if model.labels_ is None or model.results_ is None:
-        warning_messages.append("The model has not been fit yet.")
-
-    masker = getattr(model, "masker_", None)
-
-    mask_plot = None
-    if masker and masker.mask_img_:
-        fig = masker._create_figure_for_report()
-        mask_plot = figure_to_png_base64(fig)
-
-    docstring = model.__doc__
-    snippet = docstring.partition("Parameters\n    ----------\n")[0]
-
-    try:
-        design_matrices = model.design_matrices_
-    except AttributeError:
-        design_matrices = []
-    design_matrices_dict = _return_design_matrices_dict(design_matrices)
-
-    contrasts_dict = _return_contrasts_dict(design_matrices, contrasts)
-
-    statistical_maps = None
-    if contrasts:
-        statistical_maps = {}
-        statistical_maps = {
-            contrast_name: model.compute_contrast(
-                contrast_val, output_type="z_score"
-            )
-            for contrast_name, contrast_val in contrasts.items()
-        }
-
-        for contrast_name, contrast_val in contrasts.items():
-            contrast_map = model.compute_contrast(
-                contrast_val, output_type="z_score"
-            )
-            fig = experimental.plotting.plot_surf_stat_map(
-                stat_map=contrast_map,
-                hemi="left",
-                title=contrast_name,
-                colorbar=True,
-                threshold=threshold,
-            )
-            statistical_maps[contrast_name] = {
-                "stat_map_img": figure_to_png_base64(fig),
-                "contrast_img": contrasts_dict[contrast_name],
-            }
-
-    body_template_path = Path(HTML_TEMPLATE_ROOT_PATH) / "glm_report.html"
-    tpl = tempita.HTMLTemplate.from_filename(
-        str(body_template_path),
-        encoding="utf-8",
-    )
-
-    css_file_path = CSS_PATH / "masker_report.css"
-    with open(css_file_path, encoding="utf-8") as css_file:
-        css = css_file.read()
-
-    body = tpl.substitute(
-        css=css,
-        title=f"Statistical Report - {_return_model_type(model)}{title}",
-        docstring=snippet,
-        warning_messages=warning_messages,
-        design_matrices_dict=design_matrices_dict,
-        parameters=parameters,
-        contrasts_dict=contrasts_dict,
-        statistical_maps=statistical_maps,
-        cluster_table_details=cluster_table_details,
-        mask_plot=mask_plot,
-        cluster_table=None,
-    )
-
-    # revert HTML safe substitutions in CSS sections
-    body = body.replace(".pure-g &gt; div", ".pure-g > div")
-
-    head_template_path = (
-        Path(HTML_TEMPLATE_ROOT_PATH) / "surface_report_head_template.html"
-    )
-    with open(str(head_template_path)) as head_file:
-        head_tpl = Template(head_file.read())
-
-    head_css_file_path = CSS_PATH / "head.css"
-    with open(head_css_file_path, encoding="utf-8") as head_css_file:
-        head_css = head_css_file.read()
-
-    return HTMLReport(
-        body=body,
-        head_tpl=head_tpl,
-        head_values={
-            "head_css": head_css,
-            "version": __version__,
-            "page_title": (
-                "Statistical Report - " f"{_return_model_type(model)}{title}"
-            ),
-        },
-    )
-
-
-def _return_design_matrices_dict(design_matrices):
-    if not design_matrices:
-        return None
-
-    design_matrices_dict = {}
-    for dmtx_count, design_matrix in enumerate(design_matrices, start=1):
-        dmtx_plot = plot_design_matrix(design_matrix)
-        dmtx_title = f"Run {dmtx_count}"
-        if len(design_matrices) > 1:
-            plt.title(dmtx_title, y=1.025, x=-0.1)
-        dmtx_plot = _resize_plot_inches(dmtx_plot, height_change=0.3)
-        url_design_matrix_svg = _plot_to_svg(dmtx_plot)
-        # prevents sphinx-gallery & jupyter from scraping & inserting plots
-        plt.close()
-
-        design_matrices_dict[dmtx_title] = url_design_matrix_svg
-    return design_matrices_dict
-
-
-def _return_contrasts_dict(design_matrices, contrasts):
-    if not design_matrices and not contrasts:
-        return None
-
-    contrasts_dict = {}
-    for design_matrix in design_matrices:
-        for contrast_name, contrast_data in contrasts.items():
-            contrast_plot = plot_contrast_matrix(
-                contrast_data, design_matrix, colorbar=True
-            )
-            contrast_plot.set_xlabel(contrast_name)
-            contrast_plot.figure.set_figheight(2)
-            contrast_plot.figure.set_tight_layout(True)
-            url_contrast_plot_svg = _plot_to_svg(contrast_plot)
-            # prevents sphinx-gallery & jupyter
-            # from scraping & inserting plots
-            plt.close()
-            contrasts_dict[contrast_name] = url_contrast_plot_svg
-
-    return contrasts_dict
 
 
 def _return_model_type(model):
